requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.4.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.25.0
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pypdf>=3.0.0
sentence-transformers>=2.2.0
